Code/backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
markdown==3.3.3
